Start version guards for the ED loop
