+++ relays/bin-substrate/src/bridges/polkadot_bulletin/polkadot_bulletin_headers_to_bridge_hub_polkadot.rs
@@ -22,12 +22,10 @@ use crate::cli::bridge::{
 };
 
 use async_trait::async_trait;
-use relay_substrate_client::{AccountKeyPairOf, Client};
 use substrate_relay_helper::{
 	equivocation::SubstrateEquivocationDetectionPipeline,
 	finality::SubstrateFinalitySyncPipeline,
 	finality_base::{engine::Grandpa as GrandpaFinalityEngine, SubstrateFinalityPipeline},
-	TransactionParams,
 };
 
 /// Description of `PolkadotBulletin` -> `PolkadotBridgeHub` finalized headers bridge.
@@ -59,20 +57,6 @@ impl SubstrateFinalityPipeline for PolkadotBulletinFinalityToBridgeHubPolkadot {
 #[async_trait]
 impl SubstrateFinalitySyncPipeline for PolkadotBulletinFinalityToBridgeHubPolkadot {
 	type SubmitFinalityProofCallBuilder = SubmitFinalityProofCallBuilder;
-
-	async fn start_relay_guards(
-		target_client: &impl Client<Self::TargetChain>,
-		_transaction_params: &TransactionParams<AccountKeyPairOf<Self::TargetChain>>,
-		enable_version_guard: bool,
-	) -> relay_substrate_client::Result<()> {
-		if enable_version_guard {
-			relay_substrate_client::guard::abort_on_spec_version_change(
-				target_client.clone(),
-				target_client.simple_runtime_version().await?.spec_version,
-			);
-		}
-		Ok(())
-	}
 }
 
 #[async_trait]
