+++ relays/bin-substrate/src/cli/relay_headers_and_messages/relay_to_relay.rs
@@ -134,13 +134,11 @@ where
 	)> {
 		<L2R as RelayToRelayHeadersCliBridge>::Finality::start_relay_guards(
 			&self.common.right.client,
-			&self.common.right.tx_params,
 			self.common.right.client.can_start_version_guard(),
 		)
 		.await?;
 		<R2L as RelayToRelayHeadersCliBridge>::Finality::start_relay_guards(
 			&self.common.left.client,
-			&self.common.left.tx_params,
 			self.common.left.client.can_start_version_guard(),
 		)
 		.await?;
